backend/requirements.txt
flask
flask-cors
yt-dlp
openai-whisper
setuptools-rust
torch
torchvision
torchaudio
